backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.28.0
celery==5.3.4
redis==5.0.1
web3==6.11.3
aiofiles==23.2.1
python-multipart==0.0.6
langchain-openai
langgraph
mcp
langchain-mcp-adapters
uagents-adapter
hyperon
